check for u

+++ 1-number_operations.c
@@ -45,10 +45,10 @@ int print_d(va_list ar_list)
  */
 int print_u(va_list ar_list)
 {
-	unsigned long int aux_Num, count_Zero, numbers;
+	unsigned int aux_Num, count_Zero, numbers;
 	int count = 0;
 
-	numbers = va_arg(ar_list, unsigned long int);
+	numbers = va_arg(ar_list, unsigned int);
 
 	aux_Num = numbers;
 	count_Zero = 1;
